🔵 migrate publicodes related stuff to infrastructure
