+++ packages/backend/src/program/application/programService.ts
@@ -6,12 +6,13 @@ import { QuestionnaireData } from '../domain/types'
 
 import ProgramsJson from '../infrastructure/programsJson'
 import { currentDateService } from '../infrastructure/currentDate'
+import { publicodesService } from '../infrastructure/publicodes'
 
 export default class ProgramService {
   private _program: ProgramFeatures
 
   constructor() {
-    this._program = new ProgramFeatures(ProgramsJson.getInstance(), currentDateService)
+    this._program = new ProgramFeatures(ProgramsJson.getInstance(), currentDateService, publicodesService)
   }
 
   public getById(id: string): Program | undefined {
