+++ packages/backend/src/program/domain/programFeatures.ts
@@ -1,5 +1,5 @@
 import { Program } from '@tee/data/src/type/program'
-import { CurrentDateService, ProgramRepository } from './spi'
+import { CurrentDateService, ProgramRepository, RulesService } from './spi'
 import { QuestionnaireData } from './types'
 import { filterPrograms } from './filterPrograms'
 import { sortPrograms } from './sortPrograms'
@@ -8,10 +8,16 @@ import { Result } from 'true-myth'
 export default class ProgramFeatures {
   private _programRepository: ProgramRepository
   private _currentDateService: CurrentDateService | undefined
+  private _rulesService: RulesService | undefined
 
-  constructor(programRepository: ProgramRepository, currentDateService: CurrentDateService | undefined = undefined) {
+  constructor(
+    programRepository: ProgramRepository,
+    currentDateService: CurrentDateService | undefined = undefined,
+    rulesService: RulesService | undefined = undefined
+  ) {
     this._programRepository = programRepository
     this._currentDateService = currentDateService
+    this._rulesService = rulesService
   }
 
   public getById(id: string): Program | undefined {
@@ -20,10 +26,10 @@ export default class ProgramFeatures {
 
   public getFilteredBy(questionnaireData: QuestionnaireData): Result<Program[], Error> {
     const allPrograms = this._programRepository.getAll()
-    if (!this._currentDateService) {
-      return Result.err(new Error('CurrentDateService should be defined to filter programs'))
+    if (!this._currentDateService || !this._rulesService) {
+      return Result.err(new Error('currentDateService and rulesService should be defined to filter programs'))
     }
-    let filteredPrograms = filterPrograms(allPrograms, questionnaireData, this._currentDateService.get())
+    let filteredPrograms = filterPrograms(allPrograms, questionnaireData, this._currentDateService.get(), this._rulesService)
     const route = questionnaireData.questionnaire_route
     if (route) {
       filteredPrograms = filteredPrograms.map((programs) => sortPrograms(programs, route))
