+++ packages/data/tools/XL2yaml/.~lock.dispositifs.xlsx#
@@ -1,0 +1,1 @@
+,pierre,pierre-GS60-2PL,06.03.2024 10:00,file:///home/pierre/.config/libreoffice/4;
